% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt_data_dictionary.R
\docType{data}
\name{gbt_data_dictionary}
\alias{gbt_data_dictionary}
\title{Description of variables in `gbt_data_dictionary` dataset.}
\format{
A tibble with 77 rows and 5 variables:
\describe{
  \item{data}{a vector of dataframe. That is Bonds, Allocations, Key Performance Indicators,Tranches and Projects.}
  \item{ref_no}{Reference number for each variable in a dataframe.}
  \item{name}{Name of each variable as it appears in their respective dataframes.}
  \item{full_name}{Full name of each variable as it appears in their respective dataframes.}
  \item{description}{Description of name of each variable as it appears in their respective dataframes..}

}
}
\source{
\url{https://www.greenbondtransparency.com/support/resources/index-gbtp-data-en-v3.xlsx}
}
\usage{
gbt_data_dictionary
}
\description{
A summary of the `gbt_data_dictionary`
}
\keyword{datasets}
